projet_big_data/real-time-sport-events-kafka/requirements.txt
kafka-python==2.2.15
python-dotenv==1.0.1
rich==13.7.0
